Refactor framework
Switch from (what I consider) an outdated framework to a more modern
one, based on what I’m using in other applications. This looks like a
lot of work, but mostly it’s removing cruft and simplifying what is, a
simple two page application. There’s still more to do (remove
foundation, which really isn’t necessary), but for now this is enough…